Divide PolynomialM and Monomial class files
