+++ spec/poly_ruby/gbasei_spec.rb
@@ -32,7 +32,7 @@ RSpec.describe GBaseI do
 
     print "--- sample ex.4.5.5a ---\n"
     Monomial.setVarOrder(["y", "x"])
-    Monomial.setTermOrder("lex")
+    Monomial.set_term_order("lex")
     f1 = PolynomialM("2x*y-x")
     f2 = PolynomialM("3y-x^2")
     gb = GBaseI.getGBaseI([f1, f2])
@@ -40,7 +40,7 @@ RSpec.describe GBaseI do
 
     print "--- sample ex.4.5.5b ---\n"
     Monomial.setVarOrder
-    Monomial.setTermOrder("deglex")
+    Monomial.set_term_order("deglex")
     f1 = PolynomialM("3x^2y-3y*z+y")
     f2 = PolynomialM("5x^2z-8z^2")
     gb = GBaseI.getGBaseI([f1, f2])
@@ -48,7 +48,7 @@ RSpec.describe GBaseI do
 
     print "--- sample ex.4.5.5c ---\n"
     Monomial.setVarOrder
-    Monomial.setTermOrder("lex")
+    Monomial.set_term_order("lex")
     f1 = PolynomialM("6x^2+y^2")
     f2 = PolynomialM("10x^2y+2x*y")
     gb = GBaseI.getGBaseI([f1, f2])
